+++ lib/src/communication/communication_api.dart
@@ -9,7 +9,6 @@ import 'package:geiger_api/src/communication/communication_helper.dart';
 import 'package:geiger_api/src/communication/geiger_communicator.dart';
 import 'package:geiger_api/src/communication/passthrough_controller.dart';
 import 'package:geiger_api/src/communication/storage_event_handler.dart';
-import 'package:geiger_api/src/communication/storage_wrapper/owner_enforcer.dart';
 import 'package:geiger_localstorage/geiger_localstorage.dart';
 import 'package:logging/logging.dart';
 
@@ -98,6 +97,7 @@ class CommunicationApi implements GeigerApi {
   late final bool isMaster;
   late Declaration _declaration;
   Mapper? _mapper;
+  StorageController? _controller;
   String? statePath;
 
   final Map<MessageType, List<PluginListener>> _listeners =
@@ -293,11 +293,16 @@ class CommunicationApi implements GeigerApi {
   /// Obtain [StorageController] to access the storage.
   @override
   StorageController? getStorage() {
-    _mapper ??= defaultMapper;
-    if (isMaster) return GenericController(id, _mapper!.getMapper());
-    final controller = OwnerEnforcerWrapper(PassthroughController(this), id);
-    registerListener([MessageType.storageEvent], PassthroughController(this));
-    return controller;
+    if (_controller == null) {
+      _mapper ??= defaultMapper;
+      if (isMaster) {
+        _controller = GenericController(id, _mapper!.getMapper());
+      } else {
+        final passthrough = _controller = PassthroughController(this);
+        registerListener([MessageType.storageEvent], passthrough);
+      }
+    }
+    return _controller;
   }
 
   @override
@@ -383,7 +388,7 @@ class CommunicationApi implements GeigerApi {
       case MessageType.registerMenu:
         var item =
             await MenuItem.fromByteArrayStream(ByteStream(null, msg.payload));
-        menuItems[StorableString(item.menu)] = item;
+        menuItems[StorableString(item.menu.path)] = item;
         await sendMessage(Message(
             id,
             msg.sourceId,
@@ -527,10 +532,9 @@ class CommunicationApi implements GeigerApi {
   }
 
   @override
-  Future<void> registerMenu(String menu, GeigerUrl action) async {
-    MenuItem mi = MenuItem(menu, action);
+  Future<void> registerMenu(MenuItem menu) async {
     ByteSink bout = ByteSink();
-    mi.toByteArrayStream(bout);
+    menu.toByteArrayStream(bout);
     bout.close();
     await CommunicationHelper.sendAndWait(
         this,
@@ -605,5 +609,11 @@ class CommunicationApi implements GeigerApi {
   @override
   Future<void> close() async {
     await _geigerCommunicator.close();
+  }
+
+  @override
+  void authorizePlugin(PluginInformation plugin) {
+    // locally authorize the plugin
+    plugins[StorableString(plugin.toString())] = plugin;
   }
 }