+++ test/user_test.dart
@@ -143,8 +143,12 @@ Future<void> luongTests() async {
           (await getGeigerApi("", GeigerApi.masterId))!.getStorage();
       ReceivePort recv = ReceivePort();
       ReceivePort err = ReceivePort();
+      recv.listen((e) {
+        print('P: $e');
+      });
       err.listen((e) {
         print('Exception occurred');
+        print("P: $e");
         throw e;
       });
       ReceivePort ext = ReceivePort();